requirements.txt
numpy==1.16.4
matplotlib==3.1.0
pip==19.2.2
tqdm==4.32.1
GitPython==3.0.7
#git==2.23.0
pandas==1.0.1
pytest==5.2.1
scipy==1.3.1
mlflow==1.6.0
toolz==0.10.0
seaborn==0.9
scikit-learn==0.22.2.post1
requests==2.22.0
dill==0.3.1.1
Pillow==6.1
imageio==2.6.1
torchvision==0.4.0
